Day 04 - fix and solution task 2

+++ day_04/task_2/scratchcards.hpp
@@ -18,10 +18,9 @@ const bool kIsDebugOn = true;
 #define DEBUG_PRINT_TESTNAME(MSG) DEBUG_PRINT(RESET << YELLOW << MSG)
 
 using CardNumber = int;
-using Amount = int;
+using Amount = long long int;
 
 int count_guessed_numbers(const std::set<int>& winning_numbers, std::list<int> scratched_numbers);
-// int calculate_round_points(const std::set<int>& winning_numbers, std::list<int> scratched_numbers);
 std::set<int> extract_winning_numbers(std::string input);
 std::list<int>  extract_scratched_numbers(std::string input);
 int extract_card_number(std::string input);
